+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - [244](https://github.com/vegaprotocol/vegacapsule/issues/244) Improve logs collector to collect rotated files and not truncate previous logs
 - [249](https://github.com/vegaprotocol/vegacapsule/pull/249) Allow to specify resources for docker tasks
 - [210](https://github.com/vegaprotocol/vegacapsule/pull/210) Add support for Vega Visor node sets
+- [239](https://github.com/vegaprotocol/vegacapsule/issues/239) Allow to install binaries from specific release tag and add version command.
 
 ### 🐛 Fixes
 - [167](https://github.com/vegaprotocol/vegacapsule/issues/167) Fix validators filter in tendermint generator